Move signin logic to a sub-component of App
