+++ src/ui/components/App.js
@@ -1,32 +1,21 @@
 import React, { useState, useEffect } from "react";
 import { connect } from "react-redux";
 import { useApolloClient, ApolloProvider } from "@apollo/client";
-import { gql, useMutation } from "@apollo/client";
 import { useAuth0 } from "@auth0/auth0-react";
 
 import DevTools from "./DevTools";
 import Account from "./Account";
 import Loader from "./shared/Loader";
+import UserAuthentication from "./UserAuthentication";
 import { AppErrors, PopupBlockedError } from "./shared/Error";
 import SharingModal from "./shared/SharingModal";
 import { isDeployPreview } from "ui/utils/environment";
-import { setUserInBrowserPrefs } from "ui/utils/browser";
 import { selectors } from "ui/reducers";
 import { actions } from "ui/actions";
 import { hasLoadingParam } from "ui/utils/environment";
 import ResizeObserverPolyfill from "resize-observer-polyfill";
 
 import "styles.css";
-
-const CREATE_SESSION = gql`
-  mutation CreateSession($object: sessions_insert_input!) {
-    insert_sessions_one(object: $object) {
-      id
-      controller_id
-      recording_id
-    }
-  }
-`;
 
 function useGetApolloClient() {
   const [apolloClient, setApolloClient] = useState(null);
@@ -67,26 +56,8 @@ function installViewportObserver({ updateNarrowMode }) {
 }
 
 function App({ theme, recordingId, modal, updateNarrowMode, updateUser, sessionId }) {
-  const auth = useAuth0();
   const { apolloClient, consentPopupBlocked } = useGetApolloClient();
-
-  const [CreateSession] = useMutation(CREATE_SESSION);
-
-  useEffect(() => {
-    setUserInBrowserPrefs(auth.user);
-    updateUser(auth.user);
-  }, [auth.user]);
-
-  useEffect(() => {
-    if (auth.user && sessionId) {
-      const object = {
-        id: sessionId,
-        recording_id: recordingId,
-        controller_id: sessionId.split("/")[0],
-      };
-      CreateSession({ variables: { object } });
-    }
-  }, [auth.user, sessionId]);
+  const auth = useAuth0();
 
   useEffect(() => {
     document.body.parentElement.className = theme;
@@ -103,6 +74,7 @@ function App({ theme, recordingId, modal, updateNarrowMode, updateUser, sessionI
 
   return (
     <ApolloProvider client={apolloClient}>
+      <UserAuthentication />
       {recordingId ? <DevTools /> : <Account />}
       {modal?.type === "sharing" ? <SharingModal /> : null}
       <AppErrors />
